refactor: service -> api
- 폴더 명 service 대신 api 로 수정
